+++ boardmodel.h
@@ -12,28 +12,28 @@ class BoardModel
 {
 public:
     BoardModel();
-    BoardModel(bool is_rnd, size_t complexity);
+    BoardModel(bool is_rnd, quint16 complexity);
 
     QPair<bool, QVariant> is_solved() const;
-    const QVector<uint8_t>& get_board() const;
+    const QVector<quint8>& get_board() const;
 
     void set_start_board();
 
-    QPair<bool, QVariant> move(uint8_t idx);
+    QPair<bool, QVariant> move(quint8 idx);
     QPair<bool, QPair<QVariant, QVariant>> back_move();
 
 private:
-    static QPair<QVector<uint8_t>, uint8_t> _gen_board(size_t complexity);
+    static QPair<QVector<quint8>, quint8> _gen_board(quint16 complexity);
     static std::random_device _rd;
-    static uint8_t _complexity_coef;
+    static quint8 _complexity_coef;
 
-    static bool _check_solved(const QVector<uint8_t>& board);
+    static bool _check_solved(const QVector<quint8>& board);
 
-    QVector<uint8_t> _board;
-    QVector<uint8_t> _start_board;
-    QStack<QPair<uint8_t, uint8_t>> _states;
-    uint64_t _num_shifts;
-    uint8_t _nul_index;
+    QVector<quint8> _board;
+    QVector<quint8> _start_board;
+    QStack<QPair<quint8, quint8>> _states;
+    quint64 _num_shifts;
+    quint8 _nul_index;
     bool _is_solved;
 };
